# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  /** @nullable */
  username?: string | null;
  email: string;
  fullName: string;
  subscriptionTier: string;
  /** @nullable */
  trialStartDate?: string | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface SignupInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 8 */
  password: string;
  confirmPassword: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface OnboardingProfile {
  id: number;
  userId: number;
  businessName: string;
  industry: string;
  region: string;
  revenueStage: string;
  /** @nullable */
  onboardingCompletedAt?: string | null;
}

export interface OnboardingInput {
  /** @minLength 1 */
  businessName: string;
  industry: string;
  region: string;
  revenueStage: string;
}

export type FinancialModelInputs = { [key: string]: unknown };

export type FinancialModelOutputs = { [key: string]: unknown };

export interface FinancialModel {
  id: number;
  userId: number;
  industry: string;
  modelName: string;
  inputs: FinancialModelInputs;
  outputs: FinancialModelOutputs;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export type FinancialModelInputInputs = { [key: string]: unknown };

export type FinancialModelInputOutputs = { [key: string]: unknown };

export interface FinancialModelInput {
  industry: string;
  modelName: string;
  inputs: FinancialModelInputInputs;
  outputs: FinancialModelInputOutputs;
}

export type ManagementAccountOperatingExpenses = { [key: string]: unknown };

export interface ManagementAccount {
  id: number;
  userId: number;
  period: string;
  revenue: string;
  cogs: string;
  grossProfit: string;
  grossMarginPct: string;
  operatingExpenses: ManagementAccountOperatingExpenses;
  ebitda: string;
  netProfit: string;
  /** @nullable */
  createdAt?: string | null;
}

export type ManagementAccountInputOperatingExpenses = {
  [key: string]: unknown;
};

export interface ManagementAccountInput {
  period: string;
  revenue: number;
  cogs: number;
  operatingExpenses: ManagementAccountInputOperatingExpenses;
}

export interface KpiBenchmark {
  id: number;
  industry: string;
  region: string;
  metric: string;
  /** @nullable */
  p10?: string | null;
  /** @nullable */
  p25?: string | null;
  /** @nullable */
  p50?: string | null;
  /** @nullable */
  p75?: string | null;
  /** @nullable */
  p90?: string | null;
}

export interface AdminSegment {
  id: number;
  userId: number;
  industry: string;
  region: string;
  revenueStage: string;
  subscriptionTier: string;
  /** @nullable */
  signupDate?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
}

export type AdminSegmentsResponseTotalsByIndustry = { [key: string]: unknown };

export type AdminSegmentsResponseTotalsByRegion = { [key: string]: unknown };

export type AdminSegmentsResponseTotalsByTier = { [key: string]: unknown };

export type AdminSegmentsResponseTotals = {
  total: number;
  byIndustry: AdminSegmentsResponseTotalsByIndustry;
  byRegion: AdminSegmentsResponseTotalsByRegion;
  byTier: AdminSegmentsResponseTotalsByTier;
};

export interface AdminSegmentsResponse {
  segments: AdminSegment[];
  totals: AdminSegmentsResponseTotals;
}

export type GetBenchmarksParams = {
  industry: string;
  region?: string;
};