# lib/api-zod/src/generated/types/financialModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialModelInputs } from "./financialModelInputs";
import type { FinancialModelOutputs } from "./financialModelOutputs";

export interface FinancialModel {
  id: number;
  userId: number;
  industry: string;
  modelName: string;
  inputs: FinancialModelInputs;
  outputs: FinancialModelOutputs;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}