# lib/api-zod/src/generated/types/managementAccountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ManagementAccountInputOperatingExpenses } from "./managementAccountInputOperatingExpenses";

export interface ManagementAccountInput {
  period: string;
  revenue: number;
  cogs: number;
  operatingExpenses: ManagementAccountInputOperatingExpenses;
}