# lib/api-zod/src/generated/types/onboardingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingInput {
  /** @minLength 1 */
  businessName: string;
  industry: string;
  region: string;
  revenueStage: string;
}