# lib/api-zod/src/generated/types/onboardingProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingProfile {
  id: number;
  userId: number;
  businessName: string;
  industry: string;
  region: string;
  revenueStage: string;
  /** @nullable */
  onboardingCompletedAt?: string | null;
}