# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synergise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  /** @nullable */
  username?: string | null;
  email: string;
  fullName: string;
  subscriptionTier: string;
  /** @nullable */
  trialStartDate?: string | null;
  /** @nullable */
  createdAt?: string | null;
}